- update UI admin

+++ src/pages/api/posts.js
@@ -1,7 +1,8 @@
-import {SUCCESS_CODE, ERROR_CODE, SUCCESS_MESSAGE} from "../../../lib/constant";
+import { SUCCESS_CODE, ERROR_CODE, SUCCESS_MESSAGE } from "../../../lib/constant";
 import prisma from "../../../lib/prisma";
+import slugify from "slugify";
 
-const postHandler = async (req, res) => {
+const postHandler = async ( req, res ) => {
     switch (req.method) {
         case "GET":
             return getPosts(req, res);
@@ -14,7 +15,7 @@ const postHandler = async (req, res) => {
     }
 };
 
-const getPosts = async (req, res) => {
+const getPosts = async ( req, res ) => {
     try {
         const posts = await prisma.post.findMany({
             include: {
@@ -40,25 +41,31 @@ const getPosts = async (req, res) => {
             }
         });
 
-        res.status(200).json({success: true, code: SUCCESS_CODE, message: SUCCESS_MESSAGE, data: posts});
+        return res.status(200).json({ success: true, code: SUCCESS_CODE, message: SUCCESS_MESSAGE, data: posts });
     } catch (error) {
         console.log(error)
-        res.status(500).json({success: true, code: ERROR_CODE, message: error, data: []});
+        return res.status(500).json({ success: true, code: ERROR_CODE, message: error, data: [] });
     }
 }
 
-const createPost = async (req, res) => {
+const createPost = async ( req, res ) => {
     try {
         const { data } = req.body;
-        console.log(data.tags.map(tag => {
-                    return {
-                        tag_slug: tag
-                    }
-                }))
+
+        const isPostExist = await prisma.post.findFirst({
+            where: {
+                slug: slugify(data.titleEN, { lower: true })
+            }
+        });
+
+        if(isPostExist) {
+            return res.status(400).json({ success: false, code: ERROR_CODE, message: "Title already exist, Please change title English!", data: [] });
+        }
+
         const newPost = await prisma.post.create({
             data: {
                 postFormat: data.postFormat,
-                slug: data.slug,
+                slug: slugify(data.titleEN.toLowerCase()),
                 featureImg: data.featureImg,
                 date: data.date,
                 post_views: data.post_views,
@@ -70,25 +77,25 @@ const createPost = async (req, res) => {
                 trending: data.trending ? data.trending : false,
                 published: data.published ? data.published : false,
                 translations: {
-                    create: [{
+                    create: [ {
                         title: data.titleEN,
                         excerpt: data.excerptEN,
                         content: data.contentEN,
                         quoteText: data.quoteTextEN,
-                        language: {connect: {code: "en"}},
+                        language: { connect: { code: "en" } },
                     }, {
                         title: data.titleVI,
                         excerpt: data.excerptVI,
                         content: data.contentVI,
                         quoteText: data.quoteTextVI,
-                        language: {connect: {code: "vi"}},
-                    }],
+                        language: { connect: { code: "vi" } },
+                    } ],
                 },
                 author: {
-                    connect: {author_slug: data.author_slug},
+                    connect: { author_slug: data.author_slug },
                 },
                 category: {
-                    connect: {cate_slug: data.cate_slug},
+                    connect: { cate_slug: data.cate_slug },
                 },
                 tags: {
                     connect: data.tags.map(tag => {
@@ -100,16 +107,29 @@ const createPost = async (req, res) => {
             },
         });
 
-        res.status(200).json({success: true, code: SUCCESS_CODE, message: SUCCESS_MESSAGE, data: newPost});
+        return res.status(200).json({ success: true, code: SUCCESS_CODE, message: SUCCESS_MESSAGE, data: newPost });
     } catch (error) {
         console.log(error)
-        res.status(500).json({success: false, code: ERROR_CODE, message: error, data: []});
+        return res.status(500).json({ success: false, code: ERROR_CODE, message: error, data: [] });
     }
 }
 
-const updatePost = async (req, res) => {
+const updatePost = async ( req, res ) => {
     try {
         const { data } = req.body;
+
+        const isPostExist = await prisma.post.findFirst({
+            where: {
+                slug: slugify(data.titleEN, { lower: true }),
+                id: {
+                    not: data.id
+                }
+            }
+        });
+
+        if(isPostExist) {
+            return res.status(400).json({ success: false, code: ERROR_CODE, message: "Title already exist, Please change title English!", data: [] });
+        }
 
         const updatedPostTranslationsEN = await prisma.postTranslation.update({
             where: {
@@ -146,6 +166,7 @@ const updatePost = async (req, res) => {
             },
             data: {
                 postFormat: data.postFormat,
+                slug: slugify(data.titleEN.toLowerCase()),
                 featureImg: data.featureImg,
                 date: data.date,
                 post_views: data.post_views,
@@ -157,10 +178,10 @@ const updatePost = async (req, res) => {
                 trending: data.trending ? data.trending : false,
                 published: data.published ? data.published : false,
                 author: {
-                    connect: {author_slug: data.author_slug},
+                    connect: { author_slug: data.author_slug },
                 },
                 category: {
-                    connect: {cate_slug: data.cate_slug},
+                    connect: { cate_slug: data.cate_slug },
                 },
                 tags: {
                     connect: data.tags.map(tag => {
@@ -172,18 +193,18 @@ const updatePost = async (req, res) => {
             }
         });
 
-        res.status(200).json({success: true, code: SUCCESS_CODE, message: SUCCESS_MESSAGE, data: updatedPost});
+        return res.status(200).json({ success: true, code: SUCCESS_CODE, message: SUCCESS_MESSAGE, data: updatedPost });
     } catch
         (error) {
         console.log(error)
-        res.status(500).json({success: false, code: ERROR_CODE, message: error, data: []});
+        return res.status(500).json({ success: false, code: ERROR_CODE, message: error, data: [] });
     }
 }
 
-const deletePost = async (req, res) => {
+const deletePost = async ( req, res ) => {
     try {
         const { slug } = req.body;
-        const {id} = await prisma.post.findUnique({
+        const post = await prisma.post.findUnique({
             where: {
                 slug: slug
             },
@@ -192,21 +213,25 @@ const deletePost = async (req, res) => {
             }
         });
 
+        if (!post) {
+            return res.status(404).json({ success: false, code: ERROR_CODE, message: "Post not found", data: [] });
+        }
+
         const deletedPostTranslate = await prisma.postTranslation.deleteMany({
             where: {
-                postId: id
+                postId: post.id
             }
         });
         const deletedPost = await prisma.post.delete({
             where: {
-                id: id
+                id: post.id
             }
         });
 
-        res.status(200).json({success: true, code: SUCCESS_CODE, message: SUCCESS_MESSAGE, data: deletedPost});
+        return res.status(200).json({ success: true, code: SUCCESS_CODE, message: SUCCESS_MESSAGE, data: deletedPost });
     } catch (error) {
         console.log(error)
-        res.status(500).json({success: false, code: ERROR_CODE, message: error, data: []});
+        return res.status(500).json({ success: false, code: ERROR_CODE, message: error, data: [] });
     }
 }
